Refactor Game Play Screen
